+++ src/Service/Route53/src/Enum/ResourceRecordSetRegion.php
@@ -6,15 +6,11 @@ namespace AsyncAws\Route53\Enum;
  * *Latency-based resource record sets only:* The Amazon EC2 Region where you created the resource that this resource
  * record set refers to. The resource typically is an Amazon Web Services resource, such as an EC2 instance or an ELB
  * load balancer, and is referred to by an IP address or a DNS domain name, depending on the record type.
- *
- * > Although creating latency and latency alias resource record sets in a private hosted zone is allowed, it's not
- * > supported.
- *
  * When Amazon Route 53 receives a DNS query for a domain name and type for which you have created latency resource
  * record sets, Route 53 selects the latency resource record set that has the lowest latency between the end user and
  * the associated Amazon EC2 Region. Route 53 then returns the value that is associated with the selected resource
  * record set.
- * Note the following:
+ * Note the following:.
  *
  * - You can only specify one `ResourceRecord` per latency resource record set.
  * - You can only create one latency resource record set for each Amazon EC2 Region.
@@ -38,11 +34,13 @@ final class ResourceRecordSetRegion
     public const CN_NORTHWEST_1 = 'cn-northwest-1';
     public const CN_NORTH_1 = 'cn-north-1';
     public const EU_CENTRAL_1 = 'eu-central-1';
+    public const EU_CENTRAL_2 = 'eu-central-2';
     public const EU_NORTH_1 = 'eu-north-1';
     public const EU_SOUTH_1 = 'eu-south-1';
     public const EU_WEST_1 = 'eu-west-1';
     public const EU_WEST_2 = 'eu-west-2';
     public const EU_WEST_3 = 'eu-west-3';
+    public const ME_CENTRAL_1 = 'me-central-1';
     public const ME_SOUTH_1 = 'me-south-1';
     public const SA_EAST_1 = 'sa-east-1';
     public const US_EAST_1 = 'us-east-1';
@@ -66,11 +64,13 @@ final class ResourceRecordSetRegion
             self::CN_NORTHWEST_1 => true,
             self::CN_NORTH_1 => true,
             self::EU_CENTRAL_1 => true,
+            self::EU_CENTRAL_2 => true,
             self::EU_NORTH_1 => true,
             self::EU_SOUTH_1 => true,
             self::EU_WEST_1 => true,
             self::EU_WEST_2 => true,
             self::EU_WEST_3 => true,
+            self::ME_CENTRAL_1 => true,
             self::ME_SOUTH_1 => true,
             self::SA_EAST_1 => true,
             self::US_EAST_1 => true,