+++ src/Service/Route53/src/Enum/VPCRegion.php
@@ -16,6 +16,7 @@ final class VPCRegion
     public const AP_SOUTHEAST_2 = 'ap-southeast-2';
     public const AP_SOUTHEAST_3 = 'ap-southeast-3';
     public const AP_SOUTH_1 = 'ap-south-1';
+    public const AP_SOUTH_2 = 'ap-south-2';
     public const CA_CENTRAL_1 = 'ca-central-1';
     public const CN_NORTH_1 = 'cn-north-1';
     public const EU_CENTRAL_1 = 'eu-central-1';
@@ -51,6 +52,7 @@ final class VPCRegion
             self::AP_SOUTHEAST_2 => true,
             self::AP_SOUTHEAST_3 => true,
             self::AP_SOUTH_1 => true,
+            self::AP_SOUTH_2 => true,
             self::CA_CENTRAL_1 => true,
             self::CN_NORTH_1 => true,
             self::EU_CENTRAL_1 => true,