+++ src/Service/Route53/CHANGELOG.md
@@ -12,6 +12,7 @@
 - AWS api-change: Add new APIs to support Route 53 IP Based Routing
 - AWS enhancement: SDK doc update for Route 53 to update some parameters with new information.
 - AWS enhancement: Documentation updates.
+- AWS api-change: Amazon Route 53 now supports the Middle East (UAE) Region (me-central-1) for latency records, geoproximity records, and private DNS for Amazon VPCs in that region.
 
 ## 0.1.1